+++ CHANGELOG.md
@@ -1,1 +1,1 @@
-- style(leveldialog): 修正弹窗宽度及添加响应式样式 #290
+- style(App): 添加 showAuthorName 配置项并更新样式代码